+++ CHANGELOG.md
@@ -43,3 +43,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement session management with creation, validation, and refresh logic
 - Implement Authentication module with creation, token generation, and JWT encoding/decoding
 - Enhance Securial controllers and specs with error handling and authorization checks
+- Implement authentication and authorization mechanisms in Securial controllers
+- implement authenticate_admin! for actions that require admin/superuser role